refactor: remove the `collections` module
The contents in the module are moved to `manager.rs`
