--- kernel/src/process/collections/mod.rs
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-pub(super) mod process;
-pub(super) mod woken_pid;
