+++ kernel/src/process/exit.rs
@@ -2,13 +2,15 @@
 
 use x86_64::software_interrupt;
 
+use super::manager;
+
 pub(crate) fn exit_process() -> ! {
     super::set_temporary_stack_frame();
     // TODO: Call this. Currently this calling will cause a panic because the `KBox` is not mapped
     // to this process.
     // super::collections::process::remove(super::manager::getpid().into());
 
-    super::collections::woken_pid::pop();
+    manager::pop();
     cause_timer_interrupt();
 }
