+++ kernel/src/process/switch.rs
@@ -1,10 +1,9 @@
+use super::manager;
+
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use {
-    super::{
-        collections::{self, woken_pid},
-        Process,
-    },
+    super::Process,
     crate::{tests, tss},
     x86_64::{registers::control::Cr3, VirtAddr},
 };
@@ -21,12 +20,12 @@ pub(crate) fn switch() -> VirtAddr {
 }
 
 fn change_current_process() {
-    woken_pid::change_active_pid();
+    manager::change_active_pid();
 }
 
 fn switch_pml4() {
     let (_, f) = Cr3::read();
-    let pml4 = collections::process::handle_running(|p| p.pml4);
+    let pml4 = manager::handle_running(|p| p.pml4);
 
     // SAFETY: The PML4 frame is correct one and flags are unchanged.
     unsafe { Cr3::write(pml4, f) }
@@ -37,9 +36,9 @@ fn register_current_stack_frame_with_tss() {
 }
 
 fn current_stack_frame_top_addr() -> VirtAddr {
-    collections::process::handle_running(Process::stack_frame_top_addr)
+    manager::handle_running(Process::stack_frame_top_addr)
 }
 
 fn current_stack_frame_bottom_addr() -> VirtAddr {
-    collections::process::handle_running(Process::stack_frame_bottom_addr)
+    manager::handle_running(Process::stack_frame_bottom_addr)
 }
